+++ src/aioamazondevices/__init__.py
@@ -1,6 +1,6 @@
 """aioamazondevices library."""
 
-__version__ = "3.0.2"
+__version__ = "3.1.16"
 
 
 from .api import AmazonDevice, AmazonEchoApi